Refactor layout and subscription components for improved structure and functionality
- Updated import paths for AuthenticationGuard and OrganizationInitializer to reflect new component locations.
- Introduced SubscriptionGuard in the RootLayout for enhanced subscription management.
- Simplified siteId state handling in SubscribeContent by removing unnecessary state updates.
- Adjusted redirect logic to ensure users are redirected to the subscription settings if already subscribed.
- Modified PricingCard layout to improve grid structure and visual consistency.
