+++ client/src/app/components/AuthenticationGuard.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect } from "react";
 import { redirect, usePathname } from "next/navigation";
-import { userStore } from "../lib/userStore";
-import { useGetSiteIsPublic } from "../api/admin/hooks/useSites";
+import { userStore } from "../../lib/userStore";
+import { useGetSiteIsPublic } from "../../api/admin/hooks/useSites";
 
 const PUBLIC_ROUTES = ["/login", "/signup", "/invitation", "/reset-password", "/as/callback"];
