+++ client/src/app/layout.tsx
@@ -7,13 +7,14 @@ import { ThemeProvider } from "next-themes";
 import { Inter } from "next/font/google";
 import Script from "next/script";
 import { NuqsAdapter } from "nuqs/adapters/next/app";
-import { AuthenticationGuard } from "../components/AuthenticationGuard";
-import { OrganizationInitializer } from "../components/OrganizationInitializer";
+import { AuthenticationGuard } from "./components/AuthenticationGuard";
+import { OrganizationInitializer } from "./components/OrganizationInitializer";
 import { Toaster } from "../components/ui/sonner";
 import { TooltipProvider } from "../components/ui/tooltip";
 import { cn } from "../lib/utils";
 import "./globals.css";
 import { ReactScan } from "./ReactScan";
+import { SubscriptionGuard } from "./components/SubscriptionGuard";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -31,6 +32,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
           <ThemeProvider attribute="class" enableSystem={true} disableTransitionOnChange>
             <TooltipProvider>
               <QueryProvider>
+                <SubscriptionGuard />
                 <OrganizationInitializer />
                 <AuthenticationGuard />
                 {children}
