+++ client/src/app/subscribe/page.tsx
@@ -18,12 +18,12 @@ function SubscribeContent() {
   const { data: activeOrg } = authClient.useActiveOrganization();
   const router = useRouter();
 
-  const [siteId, setSiteId] = useQueryState("siteId");
+  const [siteId] = useQueryState("siteId");
 
   // Redirect if already subscribed
-  // if (subscription?.status === "active") {
-  //   router.push("/settings/organization/subscription");
-  // }
+  if (subscription?.status === "active") {
+    router.push("/settings/organization/subscription");
+  }
 
   // Get the active organization ID
   const organizationId = activeOrg?.id;
